+++ src/Server/compute/ComputeRequest.java
@@ -1,4 +1,4 @@
-package server.computeEngine;
+package server.compute;
 import user.InputConfig;
 import user.OutputConfig;
